+++ src/Query/Constraint/ForeignKey.php
@@ -16,15 +16,18 @@ final class ForeignKey
     private Column\Name $column;
     private Name $target;
     private Column\Name $reference;
+    private bool $cascadeDelete;
 
     private function __construct(
         Column\Name $column,
         Name $target,
         Column\Name $reference,
+        bool $cascadeDelete = false,
     ) {
         $this->column = $column;
         $this->target = $target;
         $this->reference = $reference;
+        $this->cascadeDelete = $cascadeDelete;
     }
 
     /**
@@ -38,12 +41,22 @@ final class ForeignKey
         return new self($column, $target, $reference);
     }
 
+    public function onDeleteCascade(): self
+    {
+        return new self(
+            $this->column,
+            $this->target,
+            $this->reference,
+            true,
+        );
+    }
+
     /**
      * @return non-empty-string
      */
     public function sql(): string
     {
-        return \sprintf(
+        $sql = \sprintf(
             'CONSTRAINT `FK_%s_%s` FOREIGN KEY (%s) REFERENCES %s(%s)',
             $this->column->toString(),
             $this->reference->toString(),
@@ -51,5 +64,10 @@ final class ForeignKey
             $this->target->sql(),
             $this->reference->sql(),
         );
+
+        return $sql.(match ($this->cascadeDelete) {
+            true => ' ON DELETE CASCADE',
+            false => '',
+        });
     }
 }